+++ packages/api/cms-api/src/builds/build-templates.resolver.ts
@@ -2,9 +2,9 @@ import { Query, Resolver } from "@nestjs/graphql";
 
 import { CurrentUserInterface } from "../auth/current-user/current-user";
 import { GetCurrentUser } from "../auth/decorators/get-current-user.decorator";
+import { KubernetesService } from "../kubernetes/kubernetes.service";
 import { BuildTemplatesService } from "./build-templates.service";
 import { BuildTemplateObject } from "./dto/build-template.object";
-import { KubernetesService } from "./kubernetes.service";
 
 @Resolver(() => BuildTemplateObject)
 export class BuildTemplatesResolver {