+++ packages/api/cms-api/src/builds/builds.resolver.ts
@@ -1,31 +1,24 @@
 import { V1CronJob } from "@kubernetes/client-node";
-import { Inject } from "@nestjs/common";
 import { Args, Mutation, Query, Resolver } from "@nestjs/graphql";
 
 import { CurrentUserInterface } from "../auth/current-user/current-user";
 import { GetCurrentUser } from "../auth/decorators/get-current-user.decorator";
 import { ContentScopeService } from "../content-scope/content-scope.service";
-import { BUILDS_CONFIG, INSTANCE_LABEL } from "./builds.constants";
-import { BuildsConfig } from "./builds.module";
+import { KubernetesService } from "../kubernetes/kubernetes.service";
+import { INSTANCE_LABEL } from "./builds.constants";
 import { BuildsService } from "./builds.service";
 import { AutoBuildStatus } from "./dto/auto-build-status.object";
 import { BuildObject } from "./dto/build.object";
 import { CreateBuildsInput } from "./dto/create-builds.input";
-import { KubernetesService } from "./kubernetes.service";
 import { SkipBuild } from "./skip-build.decorator";
 
 @Resolver(() => BuildObject)
 export class BuildsResolver {
-    helmRelease: string;
-
     constructor(
-        @Inject(BUILDS_CONFIG) readonly config: BuildsConfig,
         private readonly kubernetesService: KubernetesService,
         private readonly buildsService: BuildsService,
         private readonly contentScopeService: ContentScopeService,
-    ) {
-        this.helmRelease = config.helmRelease;
-    }
+    ) {}
 
     @Mutation(() => Boolean)
     @SkipBuild()
@@ -36,8 +29,7 @@ export class BuildsResolver {
         const cronJobs: V1CronJob[] = [];
         for (const name of names) {
             const cronJob = await this.kubernetesService.getCronJob(name);
-
-            if (this.helmRelease !== cronJob.metadata?.labels?.[INSTANCE_LABEL]) {
+            if (this.kubernetesService.helmRelease !== cronJob.metadata?.labels?.[INSTANCE_LABEL]) {
                 throw new Error("Triggering build from different instance is not allowed");
             }