+++ packages/api/cms-api/src/kubernetes/kubernetes.constants.ts
@@ -1,0 +1,2 @@
+export const KUBERNETES_MODULE_OPTIONS = "kubernetes-module-options";
+export const KUBERNETES_CONFIG = "kubernetes-config";